+++ src/actions/fileActions.ts
@@ -7,7 +7,6 @@
 /* eslint-disable import/no-cycle */
 
 import electron from 'electron';
-import Store from 'electron-store';
 import { readFile, stat, Stats, statSync } from 'fs';
 import MemoryMap, { MemoryMapTuple, Overlap } from 'nrf-intel-hex';
 import { basename } from 'path';
@@ -25,6 +24,7 @@ import {
 import { targetInfoKnown } from '../reducers/targetReducer';
 import { RootState, TDispatch } from '../reducers/types';
 import { fileWarningAdd, fileWarningRemove } from '../reducers/warningReducer';
+import { getMruFiles, setMruFiles } from '../store';
 import {
     CoreDefinition,
     deviceDefinition,
@@ -38,8 +38,6 @@ import {
     RegionName,
 } from '../util/regions';
 import { updateTargetWritable } from './targetActions';
-
-const persistentStore = new Store({ name: 'pc-nrfconnect-programmer' });
 
 export const ERROR_DIALOG_SHOW = 'ERROR_DIALOG_SHOW';
 
@@ -312,24 +310,19 @@ export const closeFiles =
     };
 
 export const loadMruFiles = () => (dispatch: TDispatch) => {
-    const files = persistentStore.get('mruFiles', []);
+    const files = getMruFiles();
     dispatch(mruFilesLoadSuccess(files));
 };
 
 const removeMruFile = (filename: string) => {
-    const files = persistentStore.get('mruFiles', []);
-    persistentStore.set(
-        'mruFiles',
-        files.filter((file: string) => file !== filename)
-    );
+    const files = getMruFiles();
+    setMruFiles(files.filter((file: string) => file !== filename));
 };
 
 const addMruFile = (filename: string) => {
-    const files = persistentStore.get('mruFiles', []);
+    const files = getMruFiles();
     if (files.indexOf(filename) === -1) {
-        files.unshift(filename);
-        files.splice(10);
-        persistentStore.set('mruFiles', files);
+        setMruFiles([filename, ...files.slice(0, 9)]);
     }
 };