+++ Changelog.md
@@ -1,3 +1,7 @@
+## Unreleased
+### Changed
+- Use shared code for persisting local settings in app.
+
 ## 2.2.0 - 2022-01-13
 ### Added
 - Auto-reset setting which, if toggled, adds a device reset after read/write operations. This setting is enabled by default.